Refactor dev menu
